Add deletion warning to TaskDelete

+++ src/components/TaskDelete.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { View, Text, TouchableOpacity } from 'react-native';
+import { View, Text, TouchableOpacity, Alert } from 'react-native';
 import { Icon } from 'react-native-elements';
 
 import { styles, fonts } from '../styles';
@@ -11,8 +11,22 @@ export default class TaskDelete extends React.Component {
   }
 
   _handleDelete() {
-    this.props.onDeleteTask(this.props.id);
-    this.props.navigation.goBack();
+    Alert.alert(
+      'Deletion Warning',
+      'Once the task is deleted, it cannot be restored. Are you sure?',
+      [
+        {
+          text: 'Cancel'
+        },
+        {
+          text: 'OK',
+          onPress: () => {
+            this.props.onDeleteTask(this.props.id);
+            this.props.navigation.goBack();
+          }
+        }
+      ]
+    );    
   }
 
   render() {
